React router introduced
App splited into smaller peaces
